% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{myNRML}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{x0}

\item{delta}{delta}

\item{llik}{llik}

\item{xrange}{xrange}

\item{parameter}{parameter}
}
\description{
myNRML
}
