% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project1Fall23.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{p}{}
}
\value{
a list of values (nd,nc,N,p,gamma). a layout with two plots (one discrete and one continuous)
}
\description{
ntickets
}
\examples{
ntickets(200,0.02,0.95)
ntickets(N=400,gamma=0.02,p=0.95)

}
